errors: create new module for errors
Move errors from backend to new errors module.
